feat: new functions to cast values to correct type
In many application repositories, we have an environmental variable
validation engine. It is typical for this engine to have code that needs
to transform incoming environmental variable values to values of a
specific type. These tasks are repeatable in many projects:
https://github.com/lidofinance/ethereum-validators-monitoring/pull/224
https://github.com/lidofinance/node-operators-widget-backend-ts/pull/41
https://github.com/lidofinance/lido-keys-api/pull/216

It makes sense to move this repeated code from many projects to some
common place.

The new "transform" module introduced in this PR collects these common
transformation functions.